refactor(studio): extract render/spatialization commands into commands/render.rs
Phase 2, second command-module extraction. Moves the 27 spatialization /
render-backend Tauri commands (spread, distance model & diffuse, hybrid, the
generic backend param setter, and the cartesian/polar evaluation settings) out of
main.rs into `commands::render`, following the pattern set in the previous PR.

No behaviour change: command names and signatures are unchanged, so the OSC
addresses and the frontend `invoke(...)` contract are identical. The shared
`send_distance_metric` helper (it sat physically among these commands) stays in
main.rs next to the other `send_*` helpers.

main.rs drops from 2523 to 2166 lines (2690 -> 2166 across the two PRs). Tauri
app builds; fmt clean.
